pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lite_llm_studio"
version = "0.0.1"
description = "Tool for fine-tuning, monitoring, and instrumenting LLMs."
readme = "README.md"
authors = [
  { name = "Lucas Bandeira" },
  { name = "Vermon Aguiar" }
]
requires-python = ">=3.12"

# Runtime dependencies only
dependencies = [
  "psutil==7.0.0",
  "pydantic==2.11.7",
  "PyYAML>=6.0",
  "streamlit>=1.49.1",
  "pyinstaller==6.15.0",
]

# Development-only dependencies
# Installed only in CI and development environments
[project.optional-dependencies]
dev = [
  "pytest>=8",
  "coverage[toml]>=7",
  "mypy>=1.10",
  "ruff>=0.5.0",
  "black>=24.0",
  "types-PyYAML==6.0.12.20250822",
]

# Optional dependencies for enhanced document processing
data = [
  "docling>=1.0.0",
  "PyPDF2>=3.0.0",
  "python-docx>=0.8.11",
]

[tool.setuptools]
package-dir = { "" = "src" }
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["lite_llm_studio*"]

# -----------------------------
# Testing and Quality
# -----------------------------
[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"
junit_family = "xunit2"

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
skip_empty = true
show_missing = true
fail_under = 0

[tool.black]
line-length = 150
target-version = ["py312"]
exclude = '''
/(
    \.git
  | \.venv
  | env
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py312"
line-length = 150
extend-exclude = ["env", ".venv", "build", "dist"]
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
mypy_path = ["src"]
